Redefine editorclass::match
It now checks the surrounding sides and corners of the block and encodes that information into a byte of data.

I also made the return value of match a new `matchcode` type, which is a typedef of `unsigned char`.

I still need to change the code to respond to this information formatting change.